requirements_no_torch.txt
numpy==1.26.4
opencv-python==4.10.0.84
tk==0.1.0
pillow==9.5.0
onnx==1.16.1
protobuf==4.23.2
tqdm
ftfy
regex
customtkinter
pyvirtualcam==0.11.1
psutil==6.0.0
onnxruntime-gpu==1.18.0
packaging==24.1
tensorrt-cu11==10.4.0
numexpr
kornia
requests